requirements.txt
pandas
numpy
requests
python-binance
google-cloud-storage
torch
pyarrow  # for parquet